Embedded test_helper into test file for ruby/ruby repo
