--- test/test_helper.rb
@@ -1,23 +1,0 @@
-# frozen_string_literal: true
-
-if /mswin|mingw|cygwin/ =~ RUBY_PLATFORM
-
-  $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
-  require "win32/registry"
-  require "test/unit"
-
-  module RegistryHelper
-    private def backslachs(path)
-      path.gsub("/", "\\")
-    end
-
-    TEST_REGISTRY_KEY = "SOFTWARE/ruby-win32-registry-test/"
-
-    def setup
-      Win32::Registry::HKEY_CURRENT_USER.open(backslachs(File.dirname(TEST_REGISTRY_KEY))) do |reg|
-        reg.delete_key File.basename(TEST_REGISTRY_KEY), true
-      end
-    rescue Win32::Registry::Error
-    end
-  end
-end
